prov/tcp: Merge tcpx_comm.c with tcpx_progress.c
comm.c only has 2 functions now, both of which are only called
from progress.c.  Move the calls there, make them static, and
remove the file.
